Rewrote module for Directions API v5
Rewrote most of the implementation for compatibility with Mapbox Directions API v5. The primary differences are an intermediate MBRouteLeg object that contains any number of MBRouteSteps, more granular properties for route steps, and geometries segmented by step.

Along the way, I also refactored the response logic and unwrapped a number of previously optional properties. It’s reasonable for this library to assert or crash if the server returns a malformed response.