Inverted default for allowUTurnAtWaypoint
The API documentation was incorrect regarding the implicit default value for the continue_straight query parameter (which depends on the profile identifier). It’s far easier to U-turn on a bike or on foot than it is in a car, so it must be the case that U-turns at waypoints are forbidden by default for driving directions only.