Use decodeObjectOfClass(es) methods to conform to NSSecureCoding
Checking the data type after the data has been unarchived is not sufficient because arbitrary code could have already been executed by that point. For decoding collections, a whitelist of classes that are okay to decode must be specified.

The RouteStepTests have changed to require secure coding when archiving/unarchiving the data. Tests for other coding classes don't yet exist, but could be considered as a future todo.