Send large payloads in POST requests
Unified RouteOptions and MatchOptions URL and URL request creation logic. Now the decision of whether to send a GET or POST request is based on the overall URL length, not based on the API endpoint.

Fixed an issue where waypoint names were left unescaped in the body of a POST request to the Map Matching API.

Added methods to get a request or request URL appropriate to the HTTP request method being used.