Throw error when decoding without options
When decoding an object from JSON without specifying a RouteOptions or MatchOptions object in the user info dictionary, throw an error instead of force-unwrapping an optional and crashing.

Ensure that waypoints in legs in routes/matches have the same names and leg-separation qualities as in the overall response and options object.