Guard against new platforms & archs for user agent
Without the `#else` clauses, any OS or CPU architecture not already listed here causes a build failure. This already occurs in Xcode 12.2b3 when it attempts to build an `arm64_32` architecture slice for watchOS, and of course Windows is not listed as an OS platform. However, the new OS names did not appear until Swift 5.2 and 5.3, and additional CPU architecture names are not always available in all toolchains (the aforementioned ARM variant is not present in Swift's source code, for example). Since adding more conditionals would just cause version-specific build failures, for now this PR just adds the fallback strings.

Note: This is a copy of the same change made by mapbox/mapbox-speech-swift#41