Accept request URL in lieu of configuration file
Replaced the --config option with a positional argument that is either a path to the JSON configuration file, as before, or the URL to a Directions or Map Matching API request.

Expand a tilde in the configuration file path when validating the arguments.

Removed a passage from the documentation that refers to a not-yet-implemented feature.